backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
black==23.9.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.1
filelock==3.12.3
flake8==6.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
identify==2.5.27
idna==3.4
itsdangerous==2.1.2
makefun==1.15.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.4.0
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
starlette==0.27.0
trio==0.21.0
typing_extensions==4.7.1
uvicorn==0.23.2
virtualenv==20.24.5
